fix: Add lockdown confirmation message & update exam mode banner (#215)
* Add immediate lock confirmation message to Lockdown

When a channel is locked immediately, the bot now sends a confirmation message indicating the channel is locked and when it will be unlocked. This improves user feedback for instant lockdown actions.
(previously it'd *think* forever)

* Update lockdown exam mode banner image URL

Replaces the old lockdown banner image URL with a new GitHub-hosted image for exam mode in the Lockdown command.